Fix for C++ attributes

+++ src/mpi/attr/comm_delete_attr.c
@@ -141,7 +141,7 @@ int MPI_Comm_delete_attr(MPI_Comm comm, int comm_keyval)
 		
 #ifdef HAVE_CXX_BINDING
 	    case MPID_LANG_CXX:
-		mpi_errno = (*MPIR_Process.cxx_call_delfn)( comm, 
+		mpi_errno = (*MPIR_Process.cxx_call_delfn)( (int)comm, 
 				     comm_keyval, p->value,
 				     keyval_ptr->extra_state, 
 		(void (*)(void)) keyval_ptr->delfn.C_CommDeleteFunction );
